Add auth-based root redirect and organize firebase lib
